+++ src/front/js/pages/datosadmons.js
@@ -20,8 +20,8 @@ export const DatosAdmon = () => {
 						<div key={index}
 							className="col-md-12 mx-auto admon">
 							<h1>{item.name} {item.lastname}</h1>
-							<button className="editor p-2 mx-3"><FaPencil/></button>
-							<button className="delete p-2 mx-3"><FaTrashCan/></button>
+							<button className="editor p-2 mx-3"><Link to={`/admon/${item.id}`}><FaPencil/></Link></button>
+							<button className="delete p-2 mx-3" onClick={()=> actions.adminDelete(item.id)}><FaTrashCan/></button>
 						</div>
 						);
